\name{BMRLSData}

\docType{data}

\alias{BMRLSData}
\alias{LSData}

\title{Data: Artificial DSGE Data}

\description{
  Artificial data generated from the log-linearised Lubik-Schorfheide (2007) model.
}

\usage{data(BMRLSData)}

\format{A matrix called `LSData' of size 200 x 5.}

\keyword{datasets}