\name{BMRMCData}

\docType{data}

\alias{BMRMCData}
\alias{bvarMCdata}

\title{Data: Artificial VAR Data}

\description{
  Artificial data generated from a bi-variate VAR(2) model.
  See section 6.1 of the vignette for details
  relating to model specification.
}

\usage{data(BMRMCData)}

\format{A matrix called `bvarMCdata' of size 100 x 2.}

\keyword{datasets}