\name{BMRVARData}

\docType{data}

\alias{BMRVARData}
\alias{USMacroData}

\title{Data: Monetary Policy VAR Data}

\description{
  This is an updated version of the Stock and Watson (2001) dataset,
  with inflation, unemployment, and the Federal Funds rate,
  from Q2 1954 to Q4 2011.
}

\usage{data(BMRVARData)}

\format{A dataframe of size 231 x 4.}

\source{Federal Reserve Economic Data}

\keyword{datasets}