\name{dsge}

\alias{dsge}
\alias{Rcpp_dsge}
\alias{Rcpp_dsge_gensys}
\alias{Rcpp_dsge_uhlig}

\title{DSGE Estimation.}

\description{
  Estimate a DSGE model.
}

\usage{
obj_gensys <- new(dsge_gensys)
obj_uhlig  <- new(dsge_uhlig)
}

\details{
  For technical details, see the accompanying vignette.
}

\author{Keith O'Hara}

\seealso{
  \code{\link{forecast.dsge}}, \code{\link{IRF.dsge}}, \code{\link{plot.dsge}}, \code{\link{states.dsge}}.
}

\examples{
\dontrun{
# see example files
}
}

\keyword{dsge}
\concept{dsge}