\name{BMR}
\alias{BMR}
\alias{BMR-package}
\docType{package}
\title{Bayesian Macroeconometrics in R}
\description{
Bayesian Macroeconometrics in R (`BMR') is a collection of R and C++ routines for estimating Bayesian Vector Autoregressive (BVAR) and Dynamic Stochastic General Equilibrium (DSGE) models in the R statistical environment.
}
\details{
\tabular{ll}{
Package: \tab BMR\cr
Type: \tab Package\cr
Version: \tab 0.4.2\cr
Date: \tab 2015-01-23\cr
License: \tab GPL-3\cr
LazyLoad: \tab Yes\cr
Depends: \tab Rcpp, RcppArmadillo, doSNOW, ggplot2, grid \cr }
}
\section{How to cite this package}{

 \preformatted{@Manual{OHara2015,
 author       = {Keith O'Hara},
 title        = {{BMR}: Bayesian Macroeconometrics in R.},
 year         = {2015},
 note         = {R package version 0.4.2.}}}
\emph{}
}
\section{License}{
BMR is licensed under the GNU General Public License (GPL) version 3. The experimental nature of this software means that it is intended for academic use only.
}
\author{Keith O'Hara}