\name{BMR}

\alias{BMR}
\alias{BMR-package}

\docType{package}

\title{Bayesian Macroeconometrics in R}

\description{
Bayesian Macroeconometrics in R (`BMR') is a collection of R and C++ routines for estimating Bayesian Vector Autoregressive (BVAR) and Dynamic Stochastic General Equilibrium (DSGE) models in the R statistical environment.
}

\details{
\tabular{ll}{
Package: \tab BMR\cr
Type: \tab Package\cr
Version: \tab 0.9.1\cr
Date: \tab 2017-09-17\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab Yes\cr
Depends: \tab Rcpp, RcppArmadillo, ggplot2, grid \cr }
}

\section{How to cite this package}{

 \preformatted{@Manual{OHara2015,
 author       = {Keith O'Hara},
 title        = {{BMR}: Bayesian Macroeconometrics in R.},
 year         = {2017},
 note         = {R package version 0.9.1.}}}
\emph{}
}

\section{License}{
BMR is licensed under the GNU General Public License (GPL) version 2, or (at your option) any later version. The experimental nature of this software implies that it is intended for academic use only.
}

\author{Keith O'Hara}