\name{dsgevar}

\alias{dsgevar}
\alias{Rcpp_dsgevar}
\alias{Rcpp_dsgevar_gensys}
\alias{Rcpp_dsgevar_uhlig}

\title{DSGE-VAR Estimation.}

\description{
  Estimate a DSGE-VAR model.
}

\usage{
obj_gensys <- new(dsgevar_gensys)
obj_uhlig  <- new(dsgevar_uhlig)
}

\details{
  For technical details, see the accompanying vignette.
}

\author{Keith O'Hara}

\seealso{
  \code{\link{forecast.dsgevar}}, \code{\link{IRF.dsgevar}}, \code{\link{plot.dsgevar}}, \code{\link{states.dsgevar}}.
}

\examples{
\dontrun{
# see example files
}
}

\keyword{dsgevar}
\concept{dsgevar}